Move tolo base entity to separate module
