+++ homeassistant/components/tolo/binary_sensor.py
@@ -9,8 +9,9 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ToloSaunaCoordinatorEntity, ToloSaunaUpdateCoordinator
+from . import ToloSaunaUpdateCoordinator
 from .const import DOMAIN
+from .entity import ToloSaunaCoordinatorEntity
 
 
 async def async_setup_entry(
