+++ homeassistant/components/tolo/button.py
@@ -8,8 +8,9 @@ from homeassistant.const import EntityCategory
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ToloSaunaCoordinatorEntity, ToloSaunaUpdateCoordinator
+from . import ToloSaunaUpdateCoordinator
 from .const import DOMAIN
+from .entity import ToloSaunaCoordinatorEntity
 
 
 async def async_setup_entry(
