+++ homeassistant/components/tolo/number.py
@@ -20,8 +20,9 @@ from homeassistant.const import EntityCategory, UnitOfTime
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
-from . import ToloSaunaCoordinatorEntity, ToloSaunaUpdateCoordinator
+from . import ToloSaunaUpdateCoordinator
 from .const import DOMAIN
+from .entity import ToloSaunaCoordinatorEntity
 
 
 @dataclass(frozen=True, kw_only=True)
